% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.Markov.Chain.R
\name{Sim.Markov.Chain}
\alias{Sim.Markov.Chain}
\title{Simulation of a finite Markov chain}
\usage{
Sim.Markov.Chain(Q, n, eta0)
}
\arguments{
\item{Q}{Transition probability matrix (r x r);}

\item{n}{length of series;}

\item{eta0}{inital value in (1,...,r).}
}
\value{
\item{x}{Simulated Markov chain}
}
\description{
This function generates a Markov chain X(1), ..., X(n) with transition matrix Q,
starting from a state eta0.
}
\examples{
Q <- matrix(c(0.8, 0.3, 0.2, 0.7),2,2) ;
sim <- Sim.Markov.Chain(Q,eta0=1,n=100)

}
\author{
Bouchra R Nasri  and Bruno N Rémillard, January 31, 2019
}
