% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sn.R
\name{Sn}
\alias{Sn}
\title{Cramer-von Mises statistic for goodness-of-fit of the null hypothesis of a univariate uniform distrubtion over [0,1]}
\usage{
Sn(U)
}
\arguments{
\item{U}{vector of pseudos-observations (apprimating uniform variates)}
}
\value{
\item{Sn}{Cramér-von Mises statistic}
}
\description{
This function computes the Cramér-von Mises statistic Sn for goodness-of-fit of the null hypothesis of a univariate uniform distrubtion over [0,1]
}
\author{
Bouchra R Nasri  and Bruno N Rémillard, January 31, 2019
}
